SPY-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-dotenv
pydantic
pytest